requirements.txt
Django==3.1.1
django-extensions==3.0.6
djangorestframework==3.11.1
factory_boy==3.0.1
